Only modify PATH if the user chose to do so.
The installation page informs the user that selecting the "Use Git from
Git Bash only" option will not modify the PATH. This patch makes it so,
even during uninstall.

Please note that this patch will *not* fix the issue right away: when
upgrading Git for Windows, the uninstaller of the *previously installed*
Git for Windows is called (and that uninstaller still has the bug we
fixed here).

This fixes https://github.com/git-for-windows/git/issues/438

[jes: fixed the syntax and the code and actually verified that
it does what the commit message claims.]

Signed-off-by: Phil Brown <brownp@stellarscience.com>
Fixed-and-tested-by: Johannes Schindelin <johannes.schindelin@gmx.de>