Only modify PATH if the user chose to do so.
Additional details: See https://github.com/git-for-windows/git/issues/438

An installation page informs the user that selecting the
"Use Git from Git Bash only" option will not modify the PATH,
which is not true. For all of the options, any existing PATH
entries that start with the current installation location are
removed. Additionally, uninstall was always removing the
installation directory, even if it was not added by the installer.

This change has not been compiled or tested.

Signed-off-by: Phil Brown <brownp@stellarscience.com>