Add a helper to submit files or URLs to VirusTotal
Git for Windows has a long track record of bogus virus scanner reports.
In fact, all of the reports turned out to be false positives (which is
not a surprise, given that the releases were always prepared on a
trusted, dedicated VM).

One blatant example is where precious development time was spent on
verifying that Git for Windows 2.5.1's `/cmd/git-gui.exe` (that was
reported as infected with a Trojan) differs from its /cmd/git.exe` (that
was reported as *uninfected*) only in the four bytes in the header that
indicate a GUI vs a console program. So obviously the virus scanner's
logic was kind of questionable, otherwise it would have flagged either
both or none.

In the process of trying to work with (or against) virus scanner
developers, the idea cropped up that signing the executables might make
a difference.

To facilitate testing this hypothesis, this script was written. It
requires an API key of https://www.virustotal.com/ (registration
required) to be put into the `$HOME/_netrc` file as

	machine api.virustotal.com
	password <your API key here>

Sadly, the test demonstrated that the PortableGit-2.6.1-64-bit.7z.exe
file was flagged by the 'Jiangmin' engine as being infected by
'Trojan/Blocker.aifm' while PortableGit-2.6.1-32-bit.7z.exe was flagged
by the 'Ad-Aware' engine with the 'Gen:TrojanHeur.FU.iK0@ayy4a0ki'
label.

For 2.6.0, it is even worse. The 32-bit portable Git is flagged by both
'Ad-Aware' (as 'Gen:Trojan.Heur.FU.iK0@aij9DHai' and 'Qihoo-360' (as
'HEUR/QVM20.1.Malware.Gen'), yet the 64-bit portable Git goes scot free.

These different findings make no sense whatsoever, given that the
portable Git installers are all *identical* self-extractors with the
`.7z` archive appended. So again, the engines should have flagged them
all, or none.

At least now we have an easy-to-use script so that future analyses might
be taking less time away from what this developer is really here for:
developing Git for Windows.

This closes https://github.com/git-for-windows/git/issues/451

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>