installer: clarify required privileges
The value 'none' is not allowed and Inno Setup falls back to 'admin'. This
is not quite what we wanted, let's change it to 'poweruser' instead.

We could consider eventually making it possible to run this as regular
user, too. When we do, we also need to consider the ramifications with
regard to existing setups.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>