installer: offer --debug-wizard-page to debug individual pages
With this option, it is much more efficient to edit and test, say,
the terminal emulator options for Git Bash:

	./release.sh --debug-wizard-page=BashTerminalPage

will build a temporary installer (excluding the files) for the sole
purpose of testing the wizard page, and start it right away.

To make the turn-around even faster, the developer can then open
`install.iss` directly in Visual Studio with Visual Studio (or Inno
Setup Studio). When everything is fine, the `install.iss` file should
be copied over the existing `install.iss.in` and selectively staged.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>