installer: get rid of install.iss.in
The idea was to interpolate constants in install.iss.in and to write the
output to install.iss. However, this is a little clumsy when one wants
to edit the files in an IDE (such as Visual Studio with the Visual
Installer extension, or Inno Setup Studio).

Let's just write the configuration into the config.iss file and include
it from install.iss instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>