upload-to-github: fix release URL to edit at the end
The idea was to give the user a convenient URL to publish the release
(because it is marked as a draft unless the release already existed,
and needs to be published). Well, that did not work too well (the
URL included the file name of the first uploaded file), so let's try
that again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>