git-extra: fix source package building
There is a convenient command to build source packages from the same
definitions as the packages were built.

However, the way we built git-extra, we did not really prepare the
sources correctly. As a consequence, a source package would not contain
*any* sources.

Let's fix that.

We need to specify all the source files explicitly. Sadly, this means we
have to put the sources back into a flat file hierarchy due to a bug in
Pacman where it interprets relative paths incorrectly unless they refer
to files in the top-level directory.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>