installer: include the NOTICE.txt only when appropriate
The NOTICE.txt file is intended only for those installers that are
also served from Microsoft's Download Central. This is done for
the versions shipped as third-party add-ons with Visual Studio, to
help with the download speed (in countries such as China, GitHub's
downloads were sometimes reported to be very slow).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>