notepad: only treat files inside .git/ specially
We want to ensure that files inside .git/ (which have LF line endings)
are edited correctly with Notepad. But we really do not have any
business meddling with other files. So let's only activate the
dos2unix/unix2dos dance, as well as the wrapping of commit messages,
when editing files inside .git/, as suggested by Congyi Wu.

This fixes https://github.com/git-for-windows/git/issues/488

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>