installer: introduce a special option for the VS add-on functionality
We will start installing Git for Windows as 3rd-party component of
Visual Studio soon. Actually, we already do that, but now it will be on
by default and the installer will also be available via the Microsoft
Download Central. This requires the special notice in accordance with
the GNU Public License, which we will now install if called in the VS
add-on mode.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>