Add a helper for common maintenance tasks
Many release engineering steps are identical between versions and can
easily be automated, such as: building the Pacman packages, installing
and uploading them, making the installer and portable Git, etc.

The main trick is to keep the two different SDKs (32-bit and 64-bit) in
sync while doing so, especially when building and installing a new MSys2
runtime (it is a little bit tricky to update that runtime, as it must
not be in use at the time, e.g. due to a running Bash).

This trick is performed by the new `please.sh` script. Being a shell
script itself, it is intended to be called from a *Git Bash*, i.e. from
an end-user Git for Windows setup.

When/if needed, the appropriate shell of the respective SDK is called.
Actually, it cannot be called directly: when calling the 64-bit SDK's
Bash from a 64-bit Git Bash, the former's MSys2 runtime realizes that it
was called from an MSys2 process and fails to acknowledge that it has a
different root directory. As a consequence, the wrong root directory
would be used. So we simply call git-cmd.exe with an overridden command,
inserting a non-MSys2 layer into the call chain.

The script expects the SDKs in the default locations (C:\git-sdk-32 and
C:\git-sdk-64). The user can override that by setting the Git config
variables windows.sdk32.path and windows.sdk64.path in their
$HOME\.gitconfig).

This commit only adds a rudimentary script, to be expanded in
functionality in future commits.

The intended usage of this script includes:

- download the script into the home directory:

	cd "$HOME"
	curl -LO \
	    https://github.com/git-for-windows/build-extra/raw/master/please.sh

- build a new mingw-w64-git package:

	./please.sh build git

- install the freshly built package:

	./please.sh install git

- build a new MSys2 runtime from Git for Windows' current sources

	./please.sh build msys2-runtime

- build all installers and portable Gits for the currently installed Git
  version

	./please.sh release

This script will be enhanced/fixed on the go and its primary beneficiary
is the script's author himself. The secondary benefit is that the bus
factor of the Git for Windows project is expected to increase a little
by documenting the release flow implicitly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>