please.sh: support building packages
This commit introduces the subcommand `build` which builds Pacman
packages. It knows to build MSys2 packages in the separate SDKs, and
MinGW packages will be built in the 64-bit SDK for both architectures.

Building the current msys2-runtime is now as easy as

	please.sh build msys2-runtime

and the mingw-w64-git package can be built using

	please.sh build git

(Since `mingw-w64-git` will be built quite often, the `git` abbreviation
is supported.)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>