please.sh: build 64-bit MinGW packages last
The common workflow is to test 64-bit Git after building and installing
the packages. For technical reasons, testing Git requires a fully built
Git in the source directory. If we build the 64-bit Git package last,
then we can reuse the build products, otherwise we have to rebuild
the 64-bit versions (because the 32-bit package's generated files
would still be there).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>