please.sh: ensure all prereqs are there for git to build
When building the mingw-w64-git package, a build requirement is the
asciidoctor-extensions package. This package is not yet provided by
MSys2, so it is served from Git for Windows' Pacman repository.

Now, Git for Windows' Pacman repository is actually *two* repositories,
one for 32-bit and one for 64-bit. Each of those repositories is seen
only from the Git for Windows SDK matching the architecture.

This is a problem when we want to build mingw-w64-git for both
architectures, so we have to jump through a little hoop to install the
32-bit asciidoctor-extensions into the 64-bit SDK (until we somehow
get MSys2 to ship that package, too).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>