installer: allow overriding defaults via command-line
In some circumstances, different default settings are more appropriate
than the ones we set, e.g. when Git for Windows is offered as a
3rd-party add-on of a larger software suite.

With this patch, it is now possible to override the defaults via the
command-line, overriding even settings from an .inf file proveded via
the /LOADINF option. The syntax of the new family of command line
options is

	/o:<key>=<value>

where <key> and <value> are identical to the corresponding .inf
settings, i.e. <key> does not contain spaces. Example:

	<installer.exe> /o:PathOption=Cmd

This closes https://github.com/git-for-windows/git/issues/575

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>