edit-git-bash: only edit the resources, do not replace them
When the installer edited the `/git-bash.exe` file to change Git Bash to
use Windows' default terminal emulator instead of MinTTY, the icon was
lost.

The problem was that this developer misread the documentation. The call
to `BeginUpdateResource()` needs to pass `FALSE` as the
`bDeleteExistingResources` parameter rather than `TRUE`: According to
the documentation, if that parameter is `TRUE`, all existing resources
will be deleted, and if the parameter is `FALSE`, existing resources are
kept unless they are specifically deleted or replaced by calls to
`UpdateResource()`. For details, see:

https://msdn.microsoft.com/en-us/library/windows/desktop/ms648030(v=vs.85).aspx

This fixes https://github.com/git-for-windows/git/issues/534

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>