please.sh: when sync'ing, give git-extra a chance to patch up
The git-extra package edits a couple of files owned by other MSys2
packages, to make up for the fact that there is no better way to
override certain configurations (such as /etc/nsswitch.conf declaring
that we want to honor the HOME environment variable).

So let's always reinstall that package last when the 'sync' command
was executed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>