Embed the MSI cabinet
Without a bundle to wrap the MSI and its cabinets, the next best option
is to embed the cabinet. Embedding the cabinet does waste disk space
on the user's system drive. However, distributing a MSI with external
cabinet over the internet (also not ideal a bundle is a better choice)
a single file is the only real option.