Execute post-install.bat at end of install.
Move the post-install.bat to an unmanaged Component (Guid="") since the
script deletes itself after running and we don't want the Windows
Installer to think the MSI is always broken. Then schedule the execution
of the post-install.bat at the very end of the installation transaction
for install and repair but not uninstall.