installer: allow configuring the signtool via a Git alias
The convenience administrative script 'please.sh' already established the
Git alias 'signtool' as a way to declare how .exe files are to be signed.

Simple example:

	[alias]
		signtool = !"/c/Progra~2/Window~1/10/bin/x64/signtool.exe \
			sign //v //f /path/to/certificate.p12 //p password"

With this, `git signtool my-program.exe` will sign the file
`my-program.exe` using the certificate in /path/to/certificate.p12 with
the password 'password'.

To dual-sign, we have to resort to a slightly more complicated method:

	[alias]
		signtool = !"/c/Progra~2/Window~1/10/bin/x64/signtool.exe \
			sign //v //f /path/to/old-certificate.p12 \
			//p password //t http://timestamp.comodoca.com \
			\\\"$@\\\" && \
			/c/Progra~2/Window~1/10/bin/x64/signtool.exe \
			sign //v //f /path/to/new-certificate.p12 \
			//p password //fd sha256 //as //td sha256 \
			//tr http://timestamp.comodoca.com/?td=sha256 \
			\\\"$@\\\""

where old-certificate.p12 refers to a SHA-1 certificate and
new-certificate.p12 refers to a SHA-2 (AKA SHA-256) certificate.

Note that due to our use of the `/as` option (to append the second
signature instead of replacing the first one), we cannot use the
signtool.exe of Windows 7.1's SDK (because it does not understand that
option) but require the signtool.exe of the Windows 10 SDK instead which
can be downloaded from:

	https://dev.windows.com/en-US/downloads/windows-10-sdk

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>