Support generating NuGet packages of Git for Windows
We steal heavily from ourselves here...

Please note that the Install.ps1 script is only called if the NuGet
package is installed inside Visual Studio (and not if installed via
nuget.exe; this is a bug in nuget.exe as of time of writing).

Therefore, the user should still call tools\post-install.bat manually
or start tools\git-bash.exe once before doing anything else, if
nuget.exe is used to install this package.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>