profile.d/aliases.sh: only set aliases for existing programs
When we alias, say, `alias node="winpty node.exe"` and `node.exe` is
actually nowhere to be found in the PATH, we are greeted with a really
obscure error:

	Error 0x2 starting node.exe

This is the way of `winpty` to say: uhm, you know, I did not really
manage to execute this program. Bu bye.

Bash has a much nicer way to state that it did not find the executable:

	bash: node: command not found

So let's not let `winpty` handle programs that are not in the PATH.

This fixes https://github.com/git-for-windows/git/issues/527
and https://github.com/git-for-windows/git/issues/491

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>