Spell MSYS2 consistently in all upper-case
Even if spelling it this way pretends that it is a four-letter acronym (it
is not, it stands for "Minimal SYStem"), it is the wish of the MSYS2
developers to spell it that way. And we owe them a lot, so we might be
just as well nice to them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>