installer: clarify the known issue with console programs
The known issue talks about interactive console programs and uses Python
as an example. Many non-Python users believed that the issue therefore did
not apply to them (e.g. because they only wanted to use OpenSSL).

This patch would have made for a nice contribution in the wake of
ihttps://github.com/git-for-windows/git/issues/660

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>