installer: avoid blurry link
On some monitors, the link to the Credential Manager looks blurred. The
reason: in InnoSetup, you can only register a mouse click action per
label, so we overlay the label with another label containing just the
hyperlink. This is tricky to get right, though, as the font metrics
differ between systems.

Instead, let's just use an appropriate number of spaces and fudge the
offset a little bit so that it looks correct enough.

Reported by Pablo Nunez Navarro.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>