installer: switch the default PATH option to 'CMD'
Many Windows users are very familiar with `cmd.exe` and not really with
the Bash at all. Requiring them to call `Git CMD` explicitly just to be
able to use Git for Windows from their favorite command-line is not
really helpful, in particular when the command-line is launched from
within other software.

So let's just switch the default to adding the `git.exe` (and gitk.exe
and git-gui.exe`) to the PATH.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>