installer: replace deprecated 'isreadme'
According to InnoSetup's documentation, the 'isreadme' flag is obsolete
(even if that has not been mentioned in InnoSetup's release notes).
Let's use the recommended way to show the release notes instead.

As a bonus, we can now modify the label to say "View Release Notes"
(instead of "View ReleaseNotes.html").

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>