Move ReleaseNotes.{css,md} and update scripts
With the addition of `render-release-notes.sh`, ReleaseNotes.md and the
stylesheet should no longer live in a specific installer's directory. This
commit moves ReleaseNotes.md and ReleaseNotes.css to the `build-extra`
directory to live with the `render-release-notes.sh` script (which going
forward, will be the only script "consuming" ReleaseNotes.md).

All installers will call upon `render-release-notes.sh` to render the
release notes for inclusion with the particular installatiotn package.

Updates to various scripts that referenced ReleaseNotes.md and/or
ReleaseNotes.css were made to reflect the new location of these files.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>