msi: move git.ico so all installers can access it
The git.ico was moved to the main `build-extra` directory so that all
installers have access to it from a common, shared location. Updated the
`install.iss` file to point to the icon's new location. Also updated the
binder arguments passed to `light.exe` in MSI's `release.sh` file--it's no
longer necessary to have to "reach into" `../installer` for various
files such as the icon and the ReleaseNotes.* files.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>