installer: add git.exe to the PATH by default, ignoring previous settings
Most users simply click through the installer and do not benefit from
our previous change that added `cmd/` to the PATH enabled by default,
because during an upgrade the previous setting (which was disabled by
default) would be replayed.

Let's skip replaying the previous setting for a couple of releases so
that users get to benefit from the new default.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>