Add ssh-pageant to package list
Add `ssh-pageant` to the list of pacman MSYS2 packages to be included in
the build. This is a drop-in replacement for ssh-agent, which allows
OpenSSH to perform key-based authentication using the PuTTY authentication
agent (Pageant).

During test builds, this increased the size of the 64-bit installer by
approximately 55 KB (29.72 vs 29.78 MB). The change in size to the
portable and SDK installers was comperable.

The home page for ssh-pageant is <https://github.com/cuviper/ssh-pageant>.

Signed-off-by: Adric Norris <landstander668@gmail.com>