Port msysGit's apply-from-gmane.sh script
Upstream Git still uses the mail-based submission system. To make it
more convenient to work with this system, we had a script to apply
patches from given GMane links in the olden Git for Windows 1.x times.

This patch ports that script (style fixes, s/msysgit/git/ because
git-for-windows does not use the mail-based submission system anymore:
we prefer Pull Requests).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>