msi: the user's core.fscache choice is configured
The MSI installer now respects the user's `core.fscache` setting during
installation and configures it in the system git configuration file for
MSYS2. The setting is also recorded in the HKLM\SOFTWARE\GitForWindows
registry key as a string value to support upgrade scenarios.

This setting will be appropriately set whether the installation is
performed during a full UI install, or a reduced UI or silent
installation. The setting is appropriately configured during repair as
well.

Since git configuration options are case-insensitive, and git Boolean
configuration settings can also be set with a 0 or 1 value, the installer
takes care to ensure that it recognizes true and false case-insensitively
and also handles 0 or 1 appropriately. Inavlid values, however, will
result in the default value being used for all installation types, which
is `true`.

This property follows the Remember Property pattern while allowing a
default value to exist for the option. More information about the
Remember Property pattern can be found at [Rob Mensching's
blog](http://robmensching.com/blog/posts/2010/5/2/the-wix-toolsets-remember-property-pattern/).
The main difference is that instead of saving any command line value
received for the property, the command line value is saved only if it
differs from the default value. The Remember Property pattern, as written,
assumes the property does not have a (default) value (which means that it
would not be authored into the MSI Property table). Since this property
has a default value, if a command line value is not provided, the property
will retain it's authored value and be saved as if it had been provided on
the command line. Then, if the user is upgrading, the registry is searched
for the property value. If the installed value differs from the default,
the property value is updated. But later, the saved "command line" value
is "restored" (which wasn't really a command line value, but the property
default value), effectively disregarding the user's prior installation
setting. By only saving the command line value if it differs from the
default, this limitation of the default implementation of the Remember
Property pattern is overcome.

You will notice that for Full UI intsalls (UILevel = 5) that if the user
has chosen to not enable this property from a previous installation or on
the command line, that the property will be set to NULL (`{}`). This is
because MSI checkboxes behave very much like HTML checkboxes in that if
there is any value whatsoever for the property backing the checkbox, the
checkbox will appear checked, even if the value specified for the property
is to disable the configuration of the feature.

At best, this is confusing for the user and it appears the installer may
be broken or not respecting their command line or previous installation
choice. They would uncheck the box which would result in the property
effectively getting set to NULL.

At worst, during the UI install, the checkbox appears checked even though
the user chose to disable the option, but they "change their mind" without
ever changing the state of the checkbox. Therefore, the property retains
its original value (false) and the feature is not configured even though
the UI indicated otherwise.

Setting the property to null, then, solves this problem.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>