msi: configure Git Credentials Manager for Windows
The MSI installer now respects the users's choice of whether or not to
enable the Git Credentials Manager for Windows during installaion and
configures it in the system git configuration file for MSYS2. The setting
is also recorded in the HKLM\SOFTWARE\GitForWindows registry key as a
DWORD value to support upgrade scenarios.

The setting will be appropriately configured whether the installation is
performed during a full UI install or a reduced UI or silent installation.
The setting is appropriately configured during repair as well.

When specifying this setting on the command line, MSI will accept
case-insensitive values of "true" and "false" as well as 0 or 1.

This setting is enabled by default if the .NET Framework 4.5.1 or above is
installed.

This property follows the Remember Property pattern while allowing a
default value to exist for the option. More information about the
Remember Property pattern can be found at [Rob Mensching's
blog](http://robmensching.com/blog/posts/2010/5/2/the-wix-toolsets-remember-property-pattern/).
The main difference is that instead of saving any command line value
received for the property, the command line value is saved only if it
differs from the default value. The Remember Property pattern, as written,
assumes the property does not have a (default) value (which means that it
would not be authored into the MSI Property table). Since this property
has a default value, if a command line value is not provided, the property
will retain it's authored value and be saved as if it had been provided on
the command line. Then, if the user is upgrading, the registry is searched
for the property value. If the installed value differs from the default,
the property value is updated. But later, the saved "command line" value
is "restored" (which wasn't really a command line value, but the property
default value), effectively disregarding the user's prior installation
setting. By only saving the command line value if it differs from the
default, this limitation of the default implementation of the Remember
Property pattern is overcome.

The configuration of this setting is a bit tricky due to some
idiosyncracies of MSI.

1. As stated above, the default value for the ENABLEGCM property is 1.

2. If the user passes 0 or "false" on the command line for this option,
this value must be saved to another property, for as the installation
progresses, ENABLEGCM may change its value, and the user's command line
choice should carry all the way through to the UI (if they're performing a
full UI install) where they can still change their mind at the last minute.

3. MSI performs a registry search (to support upgrade and repair scenarios
to retrieve previously configured installation settings). If it finds a
previous installation setting, it sets ENABLEGCM. However, DWORD's come
back as formatted strings, e.g. 0x1 ==> "#1".

4. Iff. a previous installation setting was found, and it came back as
"#1", then set ENABLEGCM=1 to enable the feature.

5. Otherwise, if on the command line the user passed "false" or 0 for
ENABLEGCM, or a previous installation disabled this feature (i.e. the
registry search returned "#0"), then set ENABLEGCM=0.

6. If the .NET Framework 4.5.1 or above is not installed, set ENABLEGCM=0.

7. If the MSI is being run as a full UI intsall (which is the most likely
scenario 9 times out of 10), and ENABLEGCM=0, then set the ENABLEGCM
property to NULL (denoted by '{}').

8. During the UI phase of setup when the user can choose to
enable/disable the feature, if the checkbox is unchecked, the property is
essentially set to NULL. In order to properly record the user's
installation choice in the registry, the GitAdditionalOptionsDlg checks to
see if ENABLEGCM <> 1, and if it's not, it's explicitly set to 0. In
addition, this checkbox is disabled if the .NET Framework 4.5.1 or
above is not installed to prevent enabling the feature when it's not
appropriate to do so.

Why is ENABLEGCM set to NULL in step 7? Because this property is used to
drive the state of a checkbox on the GitAdditionalOptionsDlg dialog during
a full UI installation.

MSI checkboxes behave similarly to HTML checkboxes. If the property
backing a checkbox has _any_ value, the checkbox will appear checked in
the UI. Only the lack of a value (i.e. null) for a property will cause the
checkbox to appear unchecked.

At best, this is confusing. The user my have indicated to disable this
feature on the command line or chosen not to enable the feature
previously. But the checkbox is checked and the user still doesn't want
this feature, so they uncheck it and everything's OK. The user just
doesn't have a great UX.

At worst, the user chose to disable this option at some point, but the
checkbox appears checked (because the property has a non-null value). The
user decides at the last minute to leave the checkbox checked and enable
the feature. The problem is, because the state of the checkbox never
changed (i.e. the user didn't uncheck and then recheck it), the value of
the property is still 0, and so the feature will not be installed, even
though the UI seemed to indicate that it would be. So, after all is said
and done, if the user chose to disable the option on the command line,
or the property has a value that is not 1, or the .NET Framework 4.5.1 is
not installed, this property is set to NULL.

Lastly, this commit includes the WiX extension in compilation and linking
for detecting the installed state of the various .NET Frameworks.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>