msi: associate .git* files with default text editor
This is an initial implementation of associating .git* files with the
default text editor.

I say "initial" because how the UI is presented for this option may
change. And if the UI changes, it may change the implementation. But as a
first pass, this works as intended so that the initial MSI will have
"feature parity" with the InnoSetup installer.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>