msi: add release.sh options & update help/usage
* Added an option to allow a user to specify the output directory location
for the installer created by this script
* Added an explicit `--help` option
* Updated the help and usage text.

If "optional packages" are specified, a variable called CUSTOM is set
to 1. This doesn't do anything right now, but we will need to know if a
particular installation of Git for Windows is a "standard installation"
(i.e. an official release) or a custom MSI created by someone who ran
`release.sh` on their own machine. It has implications for upgrading
Git for Windows when new versions are released. This whole aspect will
need further discussion--so I put the plumbing in, but there's no water
flowing yet.

The same is true for the TEST variable. I saw in the InnoSetup
installer that you can flag the installer as a "test" installer. This
is just a variant on a "custom" build--but we should change the MSI
name to reflect it's a test installer. Again, plumbing installed, no
water.

I noticed that in every other script that's run, when text is output
via the `die()` function, it's redirected to stderr, except in
`release.sh` for MSI. So I updated this version of `release.sh` to
conform to all the other scripts (e.g. `installer/release.sh` and
`../make-file-list.sh`).

Lastly, at the end of argument processing, I output information to
display the output location, version, and build type of the MSI being
created.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>