msi: sort list of files from make-file-list.sh
Because `make-file-list.sh` enumerates over packages to generate the list
of files to be included in an installer, the list of filepaths are sorted
according to the sorted order of the package names. Additionally,
"miscellaneous" files which are required are also tacked on to the end of
the list towards the end of `make-file-list.sh`.

During development, it's easier to find a particular file component in the
generated `GitComponents.wxs` file if _all_ the filepaths are sorted. This
commit adds an additional sort to ensure that the list is ordered strictly
by filepath. The list is then run through an additional call to uniq to
ensure that each filepath is listed only once. Otherwise, duplicate
components could be generated, which would cause an error when building
the MSI.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>