Add render-release-notes.sh script
This commit adds a render-release-notes.sh script which allows the
InnoSetup and MSI installers' and the NuGet package's release.sh file to
create the release notes in the same way for each distribution method.

Prior to this commit, each distribution method had their own way of
generating the Release Notes for a Git for Windows release, and the
generation of the notes for each method did not necessarily produce the
same file.

This script will provide the following benefits:
* Each distribution type will call upon a standardized way of generating
the release notes for a given Git for Windows release.
* When modifications need to be made to how release notes are generated,
those modifications occur in one place and all distributions using this
script will automatically use the latest modifications to release notes
generation.
* The script supports specifying the location relative to the output
directory of where the generated ReleaseNotes.html file expects to find
its corresponding ReleaseNotes.css CSS file.
    * This allows the expected location of the CSS file to vary depending
on the distribution type (e.g. NuGet expects the CSS file in a /content
folder).
* Allows for previewing the generating ReleaseNotes.html file.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>