msi: generate and include the Release Notes
Previously, the MSI did generate or include the installation of the
ReleaseNotes.html file and its stylesheet.

With the addition of the `render-release-notes.sh` script, this is even
more trivial to include. The MSI now knows how to generate the Release
Notes and include them en the installer to be installed with Git for
Windows.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>