msi: use CDATA for WiX conditional statements
Often times in WiX, you can specify a condition that governs whether some
action occurs (or not) during execution of an MSI. Usually, these
condition statements appear as the inner text of an XML element.

While you can use XML entity references for conditions such as "not equal"
(`&lt;&gt;`), it's much easier to read these conditions if they are placed
inside of a CDATA section instead -- especially if the conditional
statement is complex.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>