msi: install a copy of compat-bash.exe as sh.exe
The InnoSetup installer installs a copy of `compat-bash.exe` as both
`bash.exe` and `sh.exe` in the top-level `INSTALLDIR\bin` folder.

Currently, the MSI installer was only installing `compat-bash.exe` as
`bash.exe`. Also, `compat-bash.exe` was always sourced from the 'mingw64`
folder. On 32-bit architectures, however, `compat-bash.exe` should be
sourced from the `mingw32` folder. This commit sources `compat-bash.exe`
from the correct architecture for `bash.exe`. I also formatted the
`<File />` element so that it occupies only one line.

Previously, the MSI was not copying `compat-bash.exe` as `sh.exe` in the
`INSTALLFOLDER\bin`. This commit fixes that oversight and also ensures
`compat-bash.exe` is sourced from the correct folder based on the
architecture the MSI is being built for.

Signed-off-by: Craig E. Shea <craig.e.shea@gmail.com>